requirements.txt
fastapi[standard]
pydantic
langchain-openai
langserve[all]
minio
pydantic-settings